% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerCalc.R
\name{CalErrorPower}
\alias{CalErrorPower}
\title{Simulated Power Calculations for Semicontinuous data}
\usage{
CalErrorPower(prob.x, prob.y, mu.x, mu.y, sd.x, sd.y, n1, n2, simu.num, B,
  alpha = 0.05)
}
\value{
datafrome of power values
}
\description{
This function runs a simulation to determine power for various tests under 
inputted conditions
}
\examples{
CalErrorPower(.5, .7, 2, 3, 1, 1, 50, 50, 100000, 25, alpha=0.05)
}
\keyword{data}
\keyword{hypothesis}
\keyword{semicontinuous}
\keyword{test}
